added start and ending screen + jetpack animation
